reworked viewRecord, added fileTree

+++ erp_forms/public/javascripts/erp_app/desktop/applications/dynamic_forms/west_region.js
@@ -291,21 +291,19 @@ Ext.define("Compass.ErpApp.Desktop.Applications.DynamicForms.WestRegion",{
           iconCls:'icon-edit',
           handler:function(btn){
             var rec = Ext.getCmp(record.data.text).query('shared_dynamiceditablegrid').first().getSelectionModel().getSelection().first();
-            Ext.getCmp(record.data.text).editRecord(rec, record.data.text);
+            if (rec) Ext.getCmp(record.data.text).editRecord(rec, record.data.text, 'edit');
           }
         },
         {
           text: "Delete Selected Record",
           iconCls:'icon-delete',
           handler:function(btn){
-            var messageBox = Ext.MessageBox.confirm('Confirm', 'Are you sure?', 
-              function(btn){
-                if (btn == 'yes'){ 
-                  var rec = Ext.getCmp(record.data.text).query('shared_dynamiceditablegrid').first().getSelectionModel().getSelection().first();
-                  Ext.getCmp(record.data.text).deleteRecord(rec, record.data.text);
-                }
-              }
-            );                 
+            var rec = Ext.getCmp(record.data.text).query('shared_dynamiceditablegrid').first().getSelectionModel().getSelection().first();
+            if (rec){
+              var messageBox = Ext.MessageBox.confirm('Confirm', 'Are you sure?', function(btn){
+                if (btn == 'yes') Ext.getCmp(record.data.text).deleteRecord(rec, record.data.text);
+              });              
+            }
           }
         }
         ]
